Creating repos

+++ src/main/java/com/webischia/ticketmanagement/Repositories/MessageRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface MessageRepository extends CrudRepository <Message,Long>{
+public interface MessageRepository extends CrudRepository <Message,Integer>{
 
-    Optional<Message> findById(long id);
+    Optional<Message> findById(int id);
     Optional<Message> findByUserMessage(int id);
     Optional<Message> findByTicketMessage(int id);
     Optional<Message> findByMessageContext(String context); //searching by messagecontext WHERE IS THE PRIVACY ???
